requirements.txt
openai==1.3.8
mysql-connector-python
python-dotenv
PyMuPDF

